task_2_ner_image/requirements.txt
easyocr>=1.7.0
transformers>=4.21.0
torch>=2.0.0
torchvision>=0.15.0
pillow>=9.0.0
numpy>=1.21.0
pandas>=1.5.0
matplotlib>=3.5.0
scikit-learn>=1.1.0
